requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
flask==2.3.3
openai==1.3.0
slack-sdk==3.21.3
discord.py==2.3.2
requests==2.31.0
python-json-logger==2.0.7 